feat: make desktop shortcut trusted automatically
Added a few user install tasks so that the FRC vscode desktop shortcut
is marked trusted after install. This is so users can skip the the
"allow launching" step to use the desktop shortcut icon. A single
package, "gnome-terminal", is used to add all the necessary packages for
gio and dbus, which should normally be installed on typical GNOME
desktops, which is typical for Ubuntu/Pop! OS, but will probably break
compatibility with others. This isn't terribly concerning for the FRC
use-case though.